FIX remove some casting warnings

+++ src/component/forcefield/HyperReducedHexahedronFEMForceField.inl
@@ -149,7 +149,7 @@ void HyperReducedHexahedronFEMForceField<DataTypes>::addDForce (const core::Mech
 
 
     it0=this->getIndexedElements()->begin();
-    unsigned int nbElementsConsidered;
+    size_t nbElementsConsidered;
     if (!d_performECSW.getValue()){
         nbElementsConsidered = this->getIndexedElements()->size();
     }
@@ -411,7 +411,7 @@ void HyperReducedHexahedronFEMForceField<DataTypes>::addKToMatrix(const core::Me
     sofa::core::behavior::MultiMatrixAccessor::MatrixRef r = matrix->getMatrix(this->mstate);
 
     it0=this->getIndexedElements()->begin();
-    unsigned int nbElementsConsidered;
+    size_t nbElementsConsidered;
     if (!d_performECSW.getValue()){
         nbElementsConsidered = this->getIndexedElements()->size();
     }
@@ -482,7 +482,7 @@ void HyperReducedHexahedronFEMForceField<DataTypes>::draw(const core::visual::Vi
     int i;
 
     it0=this->getIndexedElements()->begin();
-    unsigned int nbElementsConsidered;
+    size_t nbElementsConsidered;
     if (!d_performECSW.getValue())
         nbElementsConsidered = this->getIndexedElements()->size();
     else
